[Transforms] Move generally applicable passes to transforms
SSA maximization and merge block insertion are generally applicable transformations and so shouldn't be nested within CFToHandshake.
