+++ lib/Transforms/CMakeLists.txt
@@ -2,6 +2,8 @@ add_circt_library(CIRCTTransforms
   FlattenMemRefs.cpp
   StripDebugInfoWithPred.cpp
   MapArithToComb.cpp
+  MaximizeSSA.cpp
+  InsertMergeBlocks.cpp
 
   ADDITIONAL_HEADER_DIRS
   ${CIRCT_MAIN_INCLUDE_DIR}/circt/Transforms
